// hr-frontend/src/lib/api/generated/model/organizationUnitDetail.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface OrganizationUnitDetail {
  id?: number;
  organizationId?: number;
  code?: string;
  name?: string;
  nameAr?: string;
  unitTypeId?: number;
  type?: string;
  typeCode?: string;
  parentUnitId?: number;
  parentUnit?: string;
  manager?: string;
  employees?: number;
  activePositions?: number;
  openPositions?: number;
  childUnits?: number;
  status?: string;
  description?: string;
  descriptionAr?: string;
  startDate?: string;
  endDate?: string;
}